requirements.txt
alembic==1.12.1
bcrypt==4.0.1
celery==5.3.6
cryptography==41.0.7
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-cache2==0.2.1
flower==2.0.1
httpx==0.25.2
itsdangerous==2.1.2
Jinja2==3.1.2
orjson==3.9.10
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
redis==4.6.0
sqladmin==0.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
ujson==5.8.0
ulid-py==1.1.0
uvicorn==0.24.0.post1
websockets==12.0
WTForms==3.1.1
